// src/main/Home.type.ts
/*
一、 泄流管（虹吸管）
Q——流量，待求值，m³/s；
H——泄流孔最小作用水头，淹没出流时为上下游水位差，自由出流时为上游水位与出口中心高程的差，m，（输入参数）；
L——管道长度，m，（输入参数）；
D——直径（对于圆形管道，取内径），m，（输入参数）；
n——糙率，默认值为0.013，可修改（经验系数）；
R——水力半径，m，（中间变量）；
A——断面面积，m²/s，（中间变量）；
g——重力加速度，m/s²（固定取值，为9.81）；
ξ——局部水头损失系数，默认值为1，可修改（经验系数）；
x——出流系数，默认为值1，可修改，淹没出流时取0（经验系数）。

*/

export type BaseItem = {
  // 唯一标识符
  id: string;
  // 名称
  name: string;
  // 数量
  count?: string;
  // 基本单位
  unit?: string;
  // 是否可编辑
  editable?: boolean;
  // 默认提示
  placeholder?: string;
  // 错误信息
  errorMessage?: string;
  // 功能描述
  description?: string;
};

export type WXieLiuGuanType = {
  // Q——流量，待求值，m³/s；
  Q: BaseItem;
  // 流速值
  V: BaseItem;
  // H——泄流孔最小作用水头，淹没出流时为上下游水位差，自由出流时为上游水位与出口中心高程的差，m，（输入参数）；
  H: BaseItem;
  // L——管道长度，m，（输入参数）；
  L: BaseItem;
  // D——直径（对于圆形管道，取内径），m，（输入参数）；
  D: BaseItem;
  // n——糙率，默认值为0.013，可修改（经验系数）；
  n: BaseItem;
  // R——水力半径，m，（中间变量）；
  R: BaseItem;
  // A——断面面积，m2，（中间变量）；
  A: BaseItem;
  // g——重力加速度，m/s²（固定取值，为9.81）；
  g: BaseItem;
  // ξ——局部水头损失系数，默认值为1，可修改（经验系数）；
  ξ: BaseItem;
  // x——出流系数，默认为值1，可修改，淹没出流时取0（经验系数）。
  x: BaseItem;
};

/*

二、 泄流闸

Q——流量，待求值，m³/s；
e——闸门开启高度，m；（输入参数）；
B——泄流孔单孔宽，m；（输入参数）；
H——自由出流时为总水头，淹没出流时为上下游水位差，m，（输入参数）；
g——重力加速度，m/s²（固定取值，为9.81）；
μ——孔流流量系数当e/H≥0.03时，不同闸形式分别用下列公式计算：
θ——闸门底缘切线与水平线的夹角（输入参数）；

*/

export enum GateType {
  type1 = " 平底平板闸门 ",
  type2 = "平底弧形闸门 ",
  type3 = "曲线型实用堰平板闸门",
  type4 = "曲线型实用堰弧形闸门",
}

export type WXieLiuZhaType = {
  // Q——流量，待求值，m³/s；
  Q: BaseItem;
  // e——闸门开启高度，m；（输入参数）；
  e: BaseItem;
  // B——泄流孔单孔宽，m；（输入参数）；
  B: BaseItem;
  // H——自由出流时为总水头，淹没出流时为上下游水位差，m，（输入参数）；
  H: BaseItem;
  // g——重力加速度，m/s²（固定取值，为9.81）；
  g: BaseItem;
  // μ——孔流流量系数当e/H≥0.03时，不同闸形式分别用下列公式计算：
  μ: BaseItem;
  // θ——闸门底缘切线与水平线的夹角（输入参数）；
  θ: BaseItem;
};

/*
 Qi——第i单机流量，待求值，m³/ s；
Ni——第i单机电功率，kW，（输入参数）；
ηi——第i单机效率，默认值为80%，可修改（输入参数）；
h——实测水头，m，（输入参数）；
g——重力加速度，m/s²（固定取值，为9.81）。
*/
export type WShengTaiType = {
  // H——水头，m，（输入参数）；
  N: BaseItem;
  // η——效率，默认值为1，可修改（经验系数）；
  η: BaseItem;
  // h——功率，待求值，kW；
  h: BaseItem;
  // g——重力加速度，m/s²（固定取值，为9.81）；
  g: BaseItem;
};

/*

Q——单机流量，待求值，m³/ s；
——单机效率，%，即（有效电功率/耗用电功率）×100，包括电动机、水泵、传动系统及水管摩阻等方面的综合效率，默认值为60%，可修改（输入参数）；
N——单机电功率，kW（输入参数）；
h——抽水扬程（抽水站上下水位差），m，当出水管口中心高于水面时，则以出水管中心高程与抽水池水体水位差计算扬程。（输入参数）；
g——重力加速度，m/s²（固定取值，为9.81）。

*/

export type WShengTaiBengType = {
  // Q——单机流量，待求值，m³/ s；
  Q: BaseItem;
  // η——效率，默认值为1，可修改（经验系数）；
  η: BaseItem;
  // h——功率，待求值，kW；
  h: BaseItem;
  // g——重力加速度，m/s²（固定取值，为9.81）；
  g: BaseItem;
  // N——单机电功率，kW（输入参数）；
  N: BaseItem;
};

/*

矩形薄壁堰

Q——流量，待求值，m³/s；
CD——流量系数，默认值为0.6，可修改（经验系数）；
b——堰口宽度，m（输入参数）；
h——总水头，m，为堰前水位与堰底的高程差，（输入参数）；
g——重力加速度，m/s²（固定取值，为9.81）。

*/

export type WCeLiuYanType1 = {
  // Q——流量，待求值，m³/s；
  Q: BaseItem;
  // CD——流量系数，默认值为0.6，可修改（经验系数）；
  CD: BaseItem;
  // b——堰口宽度，m（输入参数）；
  b: BaseItem;
  // h——总水头，m，为堰前水位与堰底的高程差，（输入参数）；
  h: BaseItem;
  // g——重力加速度，m/s²（固定取值，为9.81）；
  g: BaseItem;
};

/*

三角形薄壁堰

Q——流量，待求值，m³/s；
CD——流量系数，默认值为0.6，可修改（经验系数）；
θ——堰口角，默认值为π/2，可修改（输入参数）；
h——总水头，m，为堰前水位与堰底的高程差（输入参数）；
g——重力加速度，m/s²（固定取值，为9.81）。

*/

export type WCeLiuYanType2 = {
  // Q——流量，待求值，m³/s；
  Q: BaseItem;
  // CD——流量系数，默认值为0.6，可修改（经验系数）；
  CD: BaseItem;
  // θ——堰口角，默认值为π/2，可修改（输入参数）；
  θ: BaseItem;
  // h——总水头，m，为堰前水位与堰底的高程差（输入参数）；
  h: BaseItem;
  // g——重力加速度，m/s²（固定取值，为9.81）；
  g: BaseItem;
};

/*

锐缘矩形宽顶堰

Q——流量，待求值，m³/s；
C——流量组合系数，默认值为1，可修改（经验系数）；
h——总水头，m，为堰前水位与堰底的高程差（输入参数）；
b——堰顶宽度，m（输入参数）；
g——重力加速度，m/s²（固定取值，为9.81）。

*/

export type WCeLiuYanType3 = {
  // Q——流量，待求值，m³/s；
  Q: BaseItem;
  // C——流量组合系数，默认值为1，可修改（经验系数）；
  C: BaseItem;
  // h——总水头，m，为堰前水位与堰底的高程差（输入参数）；
  h: BaseItem;
  // b——堰顶宽度，m（输入参数）；
  b: BaseItem;
  // g——重力加速度，m/s²（固定取值，为9.81）；
  g: BaseItem;
};